area and view is moved to internal package and byte model is used instead of string model
