+++ addOns/grpc/grpc.gradle.kts
@@ -8,10 +8,6 @@ zapAddOn {
         author.set("ZAP Dev Team")
         url.set("https://www.zaproxy.org/docs/desktop/addons/grpc-support/")
     }
-
-    dependencies {
-        testImplementation(project(":testutils"))
-    }
 }
 
 crowdin {
@@ -20,4 +16,8 @@ crowdin {
         tokens.put("%messagesPath%", resourcesPath)
         tokens.put("%helpPath%", resourcesPath)
     }
+}
+
+dependencies {
+    testImplementation(project(":testutils"))
 }
