+++ addOns/grpc/src/main/resources/org/zaproxy/addon/grpc/resources/Messages.properties
@@ -1,4 +1,3 @@
 grpc.desc = Allows you to inspect, attack gRPC endpoints, and decode protobuf messages
 grpc.name = gRPC Support
-grpc.tools.menu.encdec = Protobuf Message Decoder/Encoder
-grpc.tools.menu.encdec.tooltip = Decode or encode a protobuf message
+grpc.panel.view.name = gRPC
